requirements.txt
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
pathlib>=1.0.0
argparse>=1.4.0
